Js code for camera setup(temporary)
In this updated code, the video element is set to display on the webpage, and the takePhoto() function is added to draw the current frame of the video onto the canvas, convert it to a data URL using the toDataURL() method, and create a download link for the photo. When the user clicks the download link, the photo will be downloaded as a PNG file.
Additionally, an event listener is added to the video element to continuously draw the current frame of the video onto the canvas element every 16ms. This will display the camera output on the webpage as the user is taking photos